more bug issues
i ran into a couple of bugs while working on the calculator.  i was able
 to fix them. my logic was off when it came to using decimals. When i pushed
the equal button, num1 and the operator were not being reset.  i fixed these issues
by using parseFloat, and the other by creating a separate function for the
eval operator. i am using a loop to iterate through most of the buttons, but this
is becoming troublesome. so certain buttons now have  their own event listeners and
functions.